Consolidated styles and templates and fully removed BasicsPanel.
