--- src/gui/EditorWindow/BasicsPanel/BasicsPanel.axaml.cs
@@ -1,25 +1,0 @@
-using System.ComponentModel;
-
-using Avalonia.ReactiveUI;
-
-using EVTUI.ViewModels;
-
-namespace EVTUI.Views;
-
-public partial class BasicsPanel : ReactiveUserControl<BasicsPanelViewModel>, INotifyPropertyChanged
-{
-
-    // INotifyPropertyChanged Implementation
-    new public event PropertyChangedEventHandler? PropertyChanged;
-
-    protected virtual void OnPropertyChanged(string propertyName)
-    {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-    }
-
-    public BasicsPanel()
-    {
-        InitializeComponent();
-    }
-
-}
